==> notebooks/name_resolution/1-collect_unresolved_names.ipynb <==
import pandas as pd
import os

networks_dir = "../../data/networks/all/"
network_types = ["binary", "weighted"]
network_paths = []
for nt in network_types:
    networks_subdir = f"{networks_dir}{nt}/"
    for path in os.listdir(networks_subdir):
        if path.endswith(".csv"):
            network_paths.append(f"{networks_subdir}{path}")
print(f"# networks = {len(network_paths):,}")

# plant names

species = []
for path in network_paths:
    net = pd.read_csv(path).rename(columns={"Unnamed: 0": "Plant"}).set_index("Plant")
    species += net.index.str.capitalize().tolist()
print(f"# vertices = {len(species):,}")
species = list(set(species))
print(f"# unresolved names = {len(species):,}")

pd.DataFrame({"Name": species, "Id": range(len(species))}).to_csv("../../data/name_resolution/unresolved_plant_names.csv", index=False)

# pollinator names

species = []
for path in network_paths:
    net = pd.read_csv(path).rename(columns={"Unnamed: 0": "Plant"}).set_index("Plant")
    species += net.columns.tolist()
print(f"# vertices = {len(species):,}")
species = list(set(species))
print(f"# unresolved names = {len(species):,}")

pd.DataFrame({"Name": pd.Series(species).str.capitalize(), "Id": range(len(species))}).to_csv("../../data/name_resolution/unresolved_pollinator_names.csv", index=False)
