==> notebooks/name_resolution/3-map_unchecked_to_unchecked.ipynb <==
import pandas as pd
import numpy as np
import os

ccdb_orig_path = "/groups/itay_mayrose/halabikeren/PloiDB/ccdb_all_data_171122.csv"
ccdb_new_nr_path = "/groups/itay_mayrose/halabikeren/CCDB_NR/full_resolved_names.csv"
ppn_nr_path = "/groups/itay_mayrose/halabikeren/plant_pollinator_networks/data/name_resolution/plant_nr_rwfo/processed_resolved_plant_names.csv"

ccdb_orig_nr = pd.read_csv(ccdb_orig_path).drop(["Unnamed: 0", 'parsed_n', 'reference','sporophytic', 'source'], axis=1).drop_duplicates()
ccdb_orig_nr["original_name_with_authority"] = ccdb_orig_nr.original_name
ccdb_orig_nr["sp_name"] = ccdb_orig_nr.original_name.apply(lambda n: " ".join(n.lower().split(" ")[:2]))
ccdb_orig_name_to_ploidb_name = ccdb_orig_nr.set_index("original_name")["mapped_name"].dropna().to_dict()
sp_name_to_ploidb_name = ccdb_orig_nr.set_index("sp_name")["mapped_name"].dropna().to_dict()
ploidb_names = set(ccdb_orig_nr.mapped_name.str.lower())

ccdb_orig_nr["ploidb_name_genus"] = ccdb_orig_nr.mapped_name.apply(lambda n: n.lower().split(" ")[0] if pd.notna(n) else np.nan)
ccdb_orig_nr["original_name_genus"] = ccdb_orig_nr.sp_name.apply(lambda n: n.lower().split(" ")[0])
ccdb_orig_nr["resolved_name_genus"] = ccdb_orig_nr.resolved_name.apply(lambda n: n.lower().split(" ")[0] if pd.notna(n) else np.nan)
ploidb_names_with_syn = ccdb_orig_nr.loc[(ccdb_orig_nr.ploidb_name_genus != ccdb_orig_nr.original_name_genus) & (ccdb_orig_nr.ploidb_name_genus != ccdb_orig_nr.resolved_name_genus)][["original_name", "matched_name", "resolved_name", "sp_name", "mapped_name"]].mapped_name.unique().tolist()

ccdb_new_nr = pd.read_csv(ccdb_new_nr_path, encoding='latin-1')
ccdb_new_nr["original_name"] = ccdb_new_nr["original_name"].fillna(ccdb_new_nr["name_to_resolve"])
ccdb_new_nr["original_name_with_authority"] = (ccdb_new_nr.original_name + " " + ccdb_new_nr.original_authority)
josef_input = pd.read_csv("resolved_chrome_counts_original_names.csv")

names_in_joseph_output_notin_orig_ccdb = (set(josef_input.original_name.str.lower())-set(ccdb_orig_nr.original_name.str.lower()))
named_in_orig_ccdb_notin_joseph_input = (set(ccdb_orig_nr.original_name.str.lower())-set(josef_input.original_name.str.lower()))
names_in_orig_ccdb_notin_new_ccdb = (set(ccdb_orig_nr.original_name.str.lower())-set(ccdb_new_nr.original_name.str.lower()))
print(f"# names that appear in the file josef sent for NR but not in ccdb from 17.11.22 = {len(names_in_joseph_output_notin_orig_ccdb):,}")
print(f"# names that appear in the ccdb from 17.11.22 but not file josef sent for NR = {len(named_in_orig_ccdb_notin_joseph_input):,}")
print(f"# names that appear in the ccdb from 17.11.22 but not in the new CCDB NR = {len(names_in_orig_ccdb_notin_new_ccdb):,}")

ccdb_new_nr = ccdb_new_nr.loc[ccdb_new_nr.original_name.notna()]
ccdb_new_nr.original_name = ccdb_new_nr.original_name.str.lower() 
ccdb_new_nr.resolved_name = ccdb_new_nr.resolved_name.str.lower() 
ccdb_new_nr["sp_name"] = ccdb_new_nr.original_name.apply(lambda n: " ".join(n.lower().split(" ")[:2]))
ccdb_new_nr["ploidb_name"] = ccdb_new_nr.sp_name.apply(lambda n: sp_name_to_ploidb_name.get(n, np.nan))
ccdb_new_nr["ploidb_name"] = ccdb_new_nr.apply(lambda rec: rec.ploidb_name if pd.notna(rec.ploidb_name) else ccdb_orig_name_to_ploidb_name.get(rec.original_name, np.nan), axis=1)
# complement using resolved names
ccdb_new_nr["ploidb_name"] = ccdb_new_nr["ploidb_name"].str.lower()
resolved_to_ploidb_name = ccdb_new_nr.dropna(subset=["resolved_name", "ploidb_name"]).set_index("resolved_name").ploidb_name.to_dict()

ccdb_new_nr["ploidb_name_genus"] = ccdb_new_nr.ploidb_name.apply(lambda n: n.split(" ")[0] if pd.notna(n) else np.nan)
ccdb_new_nr["original_name_genus"] = ccdb_new_nr.sp_name.apply(lambda n: n.split(" ")[0])
ccdb_new_nr["resolved_name_genus"] = ccdb_new_nr.resolved_name.apply(lambda n: n.split(" ")[0] if pd.notna(n) else np.nan)

ccdb_new_nr.to_csv(ccdb_new_nr_path, index=False)
print(f"% names in new NR that have no mapped name in the original ccdb = {np.round(ccdb_new_nr.ploidb_name.isna().sum() / ccdb_new_nr.shape[0] * 100, 2)}% ({ccdb_new_nr.ploidb_name.isna().sum():,})")

inconsistent_names = ccdb_new_nr.loc[(ccdb_new_nr.ploidb_name_genus != ccdb_new_nr.original_name_genus) & (ccdb_new_nr.ploidb_name_genus != ccdb_new_nr.resolved_name_genus) & (~ccdb_new_nr.ploidb_name.isin(ploidb_names_with_syn)) & (ccdb_new_nr.status != 'Synonym')][["original_name", "matched_name", "resolved_name", "status", "sp_name", "ploidb_name"]].drop_duplicates()
print(f"# names in ploidb that neither match the original nor the resolved names = {len(inconsistent_names.ploidb_name.unique()):,}")
inconsistent_names.groupby("status").ploidb_name.count()

inconsistent_names.query("status == 'Accepted'").sort_values("ploidb_name").drop_duplicates().head()

ccdb_orig_nr.loc[ccdb_orig_nr.mapped_name.isin(inconsistent_names.loc[inconsistent_names.status == 'Accepted', "ploidb_name"])][["original_name", "matched_name", "resolved_name", "sp_name", "mapped_name"]].sort_values("mapped_name").head()

ccdb_new_nr.original_name = ccdb_new_nr.original_name.str.lower()
ccdb_new_nr.resolved_name = ccdb_new_nr.resolved_name.str.lower()
original_to_ploidb_name = ccdb_new_nr.dropna(subset=["original_name", "ploidb_name"]).set_index("original_name")["ploidb_name"].to_dict()
resolved_to_ploidb_name = ccdb_new_nr.dropna(subset=["resolved_name", "ploidb_name"]).set_index("resolved_name")["ploidb_name"].to_dict()

ppn_resolved_names = pd.read_csv(ppn_nr_path)
ppn_resolved_names["ploidb_name"] = ppn_resolved_names.resolved_name.apply(lambda n: n if n in ploidb_names else np.nan)
ppn_resolved_names["ploidb_name"] = ppn_resolved_names.apply(lambda rec: rec.ploidb_name if pd.notna(rec.ploidb_name) else resolved_to_ploidb_name.get(rec.resolved_name, np.nan), axis=1)
ppn_resolved_names["ploidb_name"] = ppn_resolved_names.apply(lambda rec: rec.ploidb_name if pd.notna(rec.ploidb_name) else original_to_ploidb_name.get(rec.original_name, np.nan), axis=1)
ppn_resolved_names["ploidb_name"] = ppn_resolved_names.apply(lambda rec: rec.ploidb_name if pd.notna(rec.ploidb_name) else (rec.original_name if rec.original_name in ploidb_names else np.nan), axis=1)
ppn_resolved_names.ploidb_name = ppn_resolved_names.ploidb_name.apply(lambda n: n.lower() if pd.notna(n) else n)
ppn_resolved_names.to_csv(ppn_nr_path, index=False)

ppn_nr_path

ppn_resolved_names.dropna(subset=["resolved_name"]).shape[0]

ppn_resolved_names.dropna(subset=["ploidb_name"]).shape[0]

ppn_resolved_names.loc[(ppn_resolved_names.ploidb_name != ppn_resolved_names.resolved_name) & (~ppn_resolved_names.original_name.str.contains(" "))].dropna(subset=["ploidb_name"])

inconsistent_ppn_names = ppn_resolved_names.loc[(ppn_resolved_names.resolved_name != ppn_resolved_names.ploidb_name) & (ppn_resolved_names.original_name != ppn_resolved_names.ploidb_name)].dropna(subset=["ploidb_name"]).sort_values("ploidb_name")
inconsistent_ppn_names.head(n=10)

ccdb_new_nr.loc[ccdb_new_nr.ploidb_name.isin(inconsistent_ppn_names.ploidb_name)].sort_values("ploidb_name")[["original_name", "matched_name", "resolved_name", "sp_name", "ploidb_name"]].drop_duplicates().head(n=10)

ccdb_orig_nr.loc[ccdb_orig_nr.mapped_name.isin(inconsistent_ppn_names.ploidb_name)].sort_values("mapped_name")[["original_name", "matched_name", "resolved_name", "sp_name", "mapped_name"]].drop_duplicates().head(n=10)

frac_classified = ppn_resolved_names.ploidb_name.notna().sum() / ppn_resolved_names.shape[0]
frac_resolved = ppn_resolved_names.resolved_name.notna().sum() / ppn_resolved_names.shape[0]
print(f"% ploidy classifications across network plants = {np.round(frac_classified*100, 2)}% ({ppn_resolved_names.ploidb_name.notna().sum():,})")
print(f"% resolved names across network plants = {np.round(frac_resolved*100, 2)}% ({ppn_resolved_names.resolved_name.notna().sum():,})")

original_names_wo_ploidb_mapping = ppn_resolved_names.loc[ppn_resolved_names.ploidb_name.isna()].original_name.dropna().unique().tolist()
resolved_names_wo_ploidb_mapping = ppn_resolved_names.loc[ppn_resolved_names.ploidb_name.isna()].resolved_name.dropna().unique().tolist()

ccdb_orig_nr.loc[(ccdb_orig_nr.mapped_name.notna()) & (ccdb_orig_nr.original_name.str.lower().isin(original_names_wo_ploidb_mapping))]

ccdb_orig_nr.loc[(ccdb_orig_nr.mapped_name.notna()) & (ccdb_orig_nr.resolved_name.str.lower().isin(resolved_names_wo_ploidb_mapping))]

ccdb_new_nr.loc[(ccdb_new_nr.ploidb_name.notna()) & (ccdb_new_nr.original_name.str.lower().isin(original_names_wo_ploidb_mapping))]

ccdb_new_nr.loc[(ccdb_new_nr.ploidb_name.notna()) & (ccdb_new_nr.resolved_name.str.lower().isin(resolved_names_wo_ploidb_mapping))]

ccdb_new_nr.loc[ccdb_new_nr.ploidb_name == 'pouteria venosa']
