==> notebooks/name_resolution/4-asses_name_resolution.ipynb <==
import pandas as pd
import numpy as np
import os
import sys
sys.path.append("../../code/data_processing/name_resolution/")
from resolved_names_curator import ResolvedNamesCurator

# plant name resolution

unresolved_names_path=f"../../data/name_resolution/unresolved_plant_names.csv"
resolved_names_path="../../data/name_resolution/plant_nr_rwfo/resolved_plant_names.csv"
processed_resolved_names_path = "../../data/name_resolution/plant_nr_rwfo/processed_resolved_plant_names.csv"
networks_metadata_path=f"../../data/networks/all/networks_metadata.csv"

unresolved_names = pd.read_csv(unresolved_names_path)
resolved_names = pd.read_csv(resolved_names_path).rename(columns={"Coded Name": "resolved_name",
                                                                  "Original name": "original_name",
                                                                  "Matched Name": "matched_name",
                                                                  "Coded Authority": "authority"})
resolved_names.resolved_name = resolved_names.resolved_name.apply(lambda name: name.replace("_", " ") if pd.notna(name) else np.nan)
try:
    resolved_names.loc[resolved_names.resolved_name.notna(), "matched_name_wo_authority"] = resolved_names.loc[resolved_names.resolved_name.notna()][["matched_name", "authority"]].apply(lambda record: record.matched_name.replace(f" {record.authority}", "").replace(" None",""), axis=1)
except:
    pass

print(f"# unresovled names = {unresolved_names.shape[0]:,}")
print(f"# resolved names = {len(resolved_names.dropna(subset=['resolved_name']).original_name.unique()):,}")
print(f"# unique resolved names = {len(resolved_names.dropna(subset=['resolved_name']).resolved_name.unique()):,}")
print(f"% coverage by name resolution = {np.round(resolved_names.query('resolved_name.notna()').shape[0]/unresolved_names.shape[0]*100,2)}%")
print(f"% coverage of ploidb = {np.round(resolved_names.query('ploidb_name.notna()').shape[0]/unresolved_names.shape[0]*100,2)}%")

genus_names = resolved_names.loc[~resolved_names.resolved_name.str.contains(" ", na=False), "resolved_name"].tolist()
print(f"# names resolved at genus level = {len(genus_names):,}")

missing_names = resolved_names.query("resolved_name.isna()").original_name.tolist()
print(f"# missing names = {len(missing_names):,}")

resolved_names_to_doc = resolved_names[["original_name", "matched_name", "resolved_name", "ploidb_name"]]
resolved_names_to_doc.original_name = resolved_names_to_doc.original_name.str.lower()
resolved_names_to_doc.matched_name = resolved_names_to_doc.matched_name.str.lower()
resolved_names_to_doc.resolved_name = resolved_names_to_doc.resolved_name.str.lower()
resolved_names_to_doc.to_csv(processed_resolved_names_path, index=False)

# pollinators name resolution

unresolved_names_path=f"../../data/name_resolution/unresolved_pollinator_names.csv"
resolved_names_path=f"../../data/name_resolution/resolved_pollinator_names.csv"
tax_data_path = "../../data/name_resolution/itis_taxonomic_data.csv"

unresolved_names = pd.read_csv(unresolved_names_path)

resolved_names = pd.read_csv(resolved_names_path).rename(columns={"Coded Name": "resolved_name",
                                                                  "Original name": "original_name",
                                                                  "Matched Name": "matched_name",
                                                                  "Coded Authority": "authority"})
resolved_names.resolved_name = resolved_names.resolved_name.apply(lambda name: name.replace("_", " ") if pd.notna(name) else np.nan)
try:
    resolved_names.loc[resolved_names.resolved_name.notna(), "matched_name_wo_authority"] = resolved_names.loc[resolved_names.resolved_name.notna()][["matched_name", "authority"]].apply(lambda record: record.matched_name.replace(f" {record.authority}", "").replace(" None",""), axis=1)
except:
    pass

print(f"# unresovled names = {unresolved_names.shape[0]:,}")
print(f"# resolved names = {len(resolved_names.dropna(subset=['resolved_name']).original_name.unique()):,}")
print(f"# unique resolved names = {len(resolved_names.dropna(subset=['resolved_name']).resolved_name.unique()):,}")
print(f"% coverage by name resolution = {np.round(resolved_names.query('resolved_name.notna()').drop_duplicates(subset=['original_name']).shape[0]/unresolved_names.shape[0]*100,2)}%")

resolved_names.head()

resolved_names.loc[(resolved_names.resolved_name.notna()) & (resolved_names.resolved_name.str.contains(' '))].drop_duplicates("original_name").shape[0]

resolved_names.loc[(resolved_names.resolved_name.notna()) & (resolved_names.resolved_name.str.contains(' '))].drop_duplicates("original_name").shape[0] / resolved_names.drop_duplicates("original_name").shape[0]

genus_names = resolved_names.loc[~resolved_names.resolved_name.str.contains(" ", na=False), "resolved_name"].drop_duplicates().tolist()
print(f"# names resolved at genus level = {len(genus_names):,}")

resolved_names.reset_index(inplace=True)

missing_names = resolved_names.query("resolved_name.isna()").original_name.tolist()
print(f"# missing names = {len(missing_names):,}")

networks_metadata["pollinators_taxonomic_level"] = networks_metadata.pollinators.apply(lambda pollinators: get_taxonomic_level([s.replace("[", "").replace("]","").replace("'", "").strip() for s in pollinators.split("', '")]))
networks_metadata = networks_metadata.drop(["plants", "pollinators"], axis=1)
print(f"number of networks with pollinators resovled at species level")
networks_metadata.query("pollinators_taxonomic_level == 'species'").groupby("network_type")["network_index"].count()

print(f"number of networks with plants and pollinators resovled at species level")
networks_metadata.to_csv(networks_metadata_path, index=False)
networks_metadata.loc[(networks_metadata.pollinators_taxonomic_level == 'species') & (networks_metadata.plants_taxonomic_level == 'species')].groupby("network_type")["network_index"].count()

resolved_names_to_doc = resolved_names[["original_name", "matched_name", "resolved_name"]]
resolved_names_to_doc.original_name = resolved_names_to_doc.original_name.str.lower()
resolved_names_to_doc.matched_name = resolved_names_to_doc.matched_name.str.lower()
resolved_names_to_doc.resolved_name = resolved_names_to_doc.resolved_name.str.lower()

tax_data = pd.read_csv(tax_data_path)
tax_data = tax_data[["complete_name", "rank_name"]]
tax_data["complete_name"] = tax_data["complete_name"].str.lower()

resolved_names_to_doc = resolved_names_to_doc.merge(tax_data, left_on="resolved_name", right_on="complete_name", how="left")
resolved_names_to_doc.set_index("original_name", inplace=True)
resolved_names_to_doc.fillna(tax_data.set_index("complete_name")["rank_name"].to_dict(), inplace=True)
resolved_names_to_doc.reset_index(inplace=True)
resolved_names_to_doc.to_csv(resolved_names_path, index=False)

resolved_names_to_doc = pd.read_csv(resolved_names_path)
resolved_names_to_doc = resolved_names_to_doc.drop_duplicates()

print(f"# unresovled names = {unresolved_names.shape[0]:,}")
print(f"# resolved names = {len(resolved_names_to_doc.dropna(subset=['resolved_name']).original_name.unique()):,}")
print(f"# unique resolved names = {len(resolved_names_to_doc.dropna(subset=['resolved_name']).resolved_name.unique()):,}")
print(f"% coverage by name resolution = {np.round(resolved_names_to_doc.query('resolved_name.notna()').shape[0]/unresolved_names.shape[0]*100,2)}%")
print(f"% records with available rank = {np.round(resolved_names_to_doc.query('rank_name.notna()').shape[0]/unresolved_names.shape[0]*100,2)}%")

resolved_names_to_doc.loc[(resolved_names_to_doc.rank_name.notna())].shape[0] / resolved_names_to_doc.shape[0]

resolved_names_to_doc.rank_name.unique()

print(f"# unique pollinator names = {len(unresolved_names.Name.unique()):,}")

test = resolved_names_to_doc.dropna(subset=["resolved_name"])
test.groupby("rank_name")["resolved_name"].count()

## update networks pollinator indices

df = pd.read_csv(networks_metadata_path)

rank_name_to_index= {'Species': 0, 'Genus': 1, 'Subfamily': 2, 'Family': 3,  'Superfamily': 4, 'Suborder': 5, 'Order': 6, 'Subclass': 7, 'Class': 8, 'Tribe': 9, np.nan: 10}
resolved_names_to_doc["rank_index"] = resolved_names_to_doc.rank_name.replace(rank_name_to_index)

def get_pollinators_ranks(path: str):
    network = pd.read_csv(path)
    pollinators =  [s.lower() for s in set(network.columns)-{"Unnamed: 0", "Plant"}]
    resolved_pollinators_ranks = resolved_names_to_doc.loc[resolved_names_to_doc.original_name.isin(pollinators)][["rank_name", "rank_index"]].sort_values("rank_index")
    if resolved_pollinators_ranks.shape[0] == 0:
        return np.nan, 10
    rec = resolved_pollinators_ranks.iloc[0]
    return rec.rank_name, rec.rank_index
    
df[["lowest_pollinator_rank_name", "lowest_pollinator_rank_index"]] = df.apply(lambda rec: get_pollinators_ranks(rec.processed_path), axis=1, result_type="expand")

df[["network_type", "network_index", "lowest_pollinator_rank_name", "highest_pollinator_rank_name", "lowest_pollinator_rank_index", "highest_pollinator_rank_index"]].head()

df = df.drop(["highest_pollinator_rank_name", "highest_pollinator_rank_index"], axis=1).rename(columns={"lowest_pollinator_rank_name": "pollinators_rank_name",
                                                                                                        "lowest_pollinator_rank_index": "pollinators_rank_index"})
df.to_csv(networks_metadata_path, index=False)
